Solve TextureManager problems
LoadAllFromFolder still not working in windows
and I'm lost on it, seriously why does windows
use \ instead of /
yeez

Change how alpha mod and color mod is applied

+++ include/Components.h
@@ -99,7 +99,7 @@ struct RenderComponent {
     float size{1.0f}; //Percentage, 1 is original size
     bool isStretched {false};
     FPair exactSize{1.0f, 1.0f};
-    int visibility{255};
+    Uint8 visibility{255};
     bool free {false};
     const SDL_Color* color {nullptr};
