+++ include/texture.hpp
@@ -35,7 +35,7 @@ public:
 		SDL_FlipMode t_flip, float t_scale = 1.0f, const SDL_Color* color = nullptr);
 	void renderEx(int t_x, int t_y, SDL_FRect* t_clip, double t_angle, SDL_FPoint* t_center, 
 		SDL_FlipMode t_flip, FPair size, const SDL_Color* color = nullptr);
-	void setAlphaMod(int alpha);
+	void setAlphaMod(Uint8 alpha);
 	void colorMod(const SDL_Color& color);
 
 	float getWidth();
